requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
streamlit==1.32.2
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
timm==0.9.2
pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
altair==5.5.0
